python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crforge-gym"
version = "0.1.0"
description = "Gymnasium environment for CRForge Clash Royale simulator"
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "pyzmq>=25.0.0",
]

[project.optional-dependencies]
train = [
    "stable-baselines3>=2.3.0",
    "sb3-contrib>=2.3.0",
    "tensorboard>=2.14.0",
]
jpype = [
    "jpype1>=1.5.0",
]
dev = [
    "stable-baselines3>=2.3.0",
    "sb3-contrib>=2.3.0",
    "tensorboard>=2.14.0",
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["crforge_gym*"]
